+++ src/app/ModalRoutes.svelte
@@ -7,6 +7,7 @@
   import LoginPubKey from "src/app/views/LoginPubKey.svelte"
   import Onboarding from "src/app/views/Onboarding.svelte"
   import NoteCreate from "src/app/views/NoteCreate.svelte"
+  import NoteDelete from "src/app/views/NoteDelete.svelte"
   import NoteZap from "src/app/views/NoteZap.svelte"
   import NoteShare from "src/app/views/NoteShare.svelte"
   import NoteDetail from "src/app/views/NoteDetail.svelte"
@@ -29,7 +30,9 @@
     <NoteDetail {...m} invertColors />
   {/key}
 {:else if m.type === "note/create"}
-  <NoteCreate pubkey={m.pubkey} nevent={m.nevent} writeTo={m.relays} />
+  <NoteCreate pubkey={m.pubkey} quote={m.quote} writeTo={m.relays} />
+{:else if m.type === "note/delete"}
+  <NoteDelete note={m.note} />
 {:else if m.type === "note/zap"}
   <NoteZap note={m.note} />
 {:else if m.type === "note/share"}