+++ src/app/shared/PersonNotes.svelte
@@ -6,7 +6,7 @@
   export let relays
   export let invertColors = false
 
-  const filter = {kinds: [1], authors: [pubkey]}
+  const filter = {kinds: [1, 1985], authors: [pubkey]}
 </script>
 
 <Feed {relays} {filter} {invertColors} parentsTimeout={3000} delta={timedelta(1, "days")} />